Extract LOB and XML tests in their own file too
No changes, just move the tests to a separate file.
