+++ tests/CMakeLists.txt
@@ -30,6 +30,7 @@ add_library(soci_tests_common STATIC
   common/test-common.cpp
   common/test-custom.cpp
   common/test-dynamic.cpp
+  common/test-lob.cpp
   test-assert.h
   test-context.h
   test-myint.h)
